litpe

+++ mystic-tales-podcast-backend/subscription-service/SubscriptionService.BusinessLogic/Services/DbServices/SubscriptionServices/PodcastSubscriptionService.cs
@@ -2313,38 +2313,55 @@ namespace SubscriptionService.BusinessLogic.Services.DbServices.SubscriptionServ
                     var flowName = command.FlowName;
                     var responseData = command.LastStepResponseData;
 
-                    var podcastSubscription = await _podcastSubscriptionGenericRepository.FindAll()
+                    var show = await GetPodcastShow(parameter.PodcastShowId);
+                    var showPodcastSubscription = await _podcastSubscriptionGenericRepository.FindAll()
                         .Include(ps => ps.PodcastSubscriptionCycleTypePrices)
                         .Where(ps => ps.PodcastShowId == parameter.PodcastShowId && ps.IsActive && ps.DeletedAt == null)
                         .FirstOrDefaultAsync();
 
+                    var channelPodcastSubscription = await _podcastSubscriptionGenericRepository.FindAll()
+                        .Include(ps => ps.PodcastSubscriptionCycleTypePrices)
+                        .Where(ps => ps.PodcastChannelId == show.PodcastChannelId && ps.IsActive && ps.DeletedAt == null)
+                        .FirstOrDefaultAsync();
+
                     var subscriptionRegistrations = await _podcastSubscriptionRegistrationGenericRepository.FindAll()
-                        .Where(sr => sr.PodcastSubscriptionId == podcastSubscription.Id && sr.CancelledAt == null)
+                        .Where(sr => sr.PodcastSubscriptionId == showPodcastSubscription.Id && sr.CancelledAt == null)
                         .ToListAsync();
+
+                    var subscriptionRegistrationsChannel = await _podcastSubscriptionRegistrationGenericRepository.FindAll()
+                        .Where(sr => sr.PodcastSubscriptionId == channelPodcastSubscription.Id && sr.CancelledAt == null)
+                        .ToListAsync();
+
                     foreach (var registration in subscriptionRegistrations)
                     {
-                        if (!registration.IsIncomeTaken)
+                        foreach (var channelRegistration in subscriptionRegistrationsChannel)
                         {
-                            var amount = podcastSubscription.PodcastSubscriptionCycleTypePrices
-                                .Where(ptcp => ptcp.SubscriptionCycleTypeId == registration.SubscriptionCycleTypeId)
-                                .Select(ptcp => ptcp.Price)
-                                .FirstOrDefault();
-                            var tempRequestData = new JObject
+                            if(registration.AccountId == channelRegistration.AccountId)
                             {
-                                { "PodcastSubscriptionRegistrationId", registration.Id },
-                                { "Profit", null },
-                                { "AccountId", registration.AccountId },
-                                { "Amount", amount },
-                                { "TransactionTypeId", (int)TransactionTypeEnum.CustomerSubscriptionCyclePaymentRefund }
-                            };
-                            var refundMessage = _kafkaProducerService.PrepareStartSagaTriggerMessage(
-                                topic: KafkaTopicEnum.PaymentProcessingDomain,
-                                requestData: tempRequestData,
-                                sagaInstanceId: null,
-                                messageName: "podcast-subscription-refund-flow");
-                            await _messagingService.SendSagaMessageAsync(refundMessage, null);
+                                if (!registration.IsIncomeTaken)
+                                {
+                                    var amount = showPodcastSubscription.PodcastSubscriptionCycleTypePrices
+                                        .Where(ptcp => ptcp.SubscriptionCycleTypeId == registration.SubscriptionCycleTypeId)
+                                        .Select(ptcp => ptcp.Price)
+                                        .FirstOrDefault();
+                                    var tempRequestData = new JObject
+                                    {
+                                        { "PodcastSubscriptionRegistrationId", registration.Id },
+                                        { "Profit", null },
+                                        { "AccountId", registration.AccountId },
+                                        { "Amount", amount },
+                                        { "TransactionTypeId", (int)TransactionTypeEnum.CustomerSubscriptionCyclePaymentRefund }
+                                    };
+                                    var refundMessage = _kafkaProducerService.PrepareStartSagaTriggerMessage(
+                                        topic: KafkaTopicEnum.PaymentProcessingDomain,
+                                        requestData: tempRequestData,
+                                        sagaInstanceId: null,
+                                        messageName: "podcast-subscription-refund-flow");
+                                    await _messagingService.SendSagaMessageAsync(refundMessage, null);
+                                }
+                                await _podcastSubscriptionRegistrationGenericRepository.DeleteAsync(registration.Id);
+                            }
                         }
-                        await _podcastSubscriptionRegistrationGenericRepository.DeleteAsync(registration.Id);
                     }
 
                     await transaction.CommitAsync();
@@ -2762,7 +2779,7 @@ namespace SubscriptionService.BusinessLogic.Services.DbServices.SubscriptionServ
             var episode = await GetPodcastEpisode(podcastEpisodeId);
             if (episode == null)
             {
-                return (false, $"Podcast episode with Id: {podcastEpisodeId} is not found");
+                return (false, $"Podcast episode is not found");
             }
             var (isValid, errorMessage) = await ValidateShow(episode.PodcastShowId, podcastEpisodeId);
             if (!isValid)
@@ -2771,28 +2788,28 @@ namespace SubscriptionService.BusinessLogic.Services.DbServices.SubscriptionServ
             }
             if (episode.DeletedAt != null)
             {
-                return (false, $"Podcast episode with Id: {podcastEpisodeId} has already been deleted");
+                return (false, $"Podcast episode has already been deleted");
             }
             var episodeStatusId = episode.PodcastEpisodeStatusTrackings.OrderByDescending(es => es.CreatedAt).Select(es => es.PodcastEpisodeStatusId).First();
             if (episodeStatusId == (int)PodcastEpisodeStatusEnum.Draft)
             {
-                return (false, $"Podcast episode with Id: {podcastEpisodeId} is in Draft status");
+                return (false, $"Podcast episode is in Draft status");
             }
             if (episodeStatusId == (int)PodcastEpisodeStatusEnum.PendingReview)
             {
-                return (false, $"Podcast episode with Id: {podcastEpisodeId} is pending review");
+                return (false, $"Podcast episode is pending review");
             }
             if (episodeStatusId == (int)PodcastEpisodeStatusEnum.PendingEditRequired)
             {
-                return (false, $"Podcast episode with Id: {podcastEpisodeId} is pending edit required");
+                return (false, $"Podcast episode is pending edit required");
             }
             if (episodeStatusId == (int)PodcastEpisodeStatusEnum.TakenDown)
             {
-                return (false, $"Podcast episode with Id: {podcastEpisodeId} has been taken down");
+                return (false, $"Podcast episode has been taken down");
             }
             if (episodeStatusId == (int)PodcastEpisodeStatusEnum.Removed)
             {
-                return (false, $"Podcast episode with Id: {podcastEpisodeId} has been removed");
+                return (false, $"Podcast episode has been removed");
             }
             return (true, string.Empty);
         }
@@ -2802,7 +2819,7 @@ namespace SubscriptionService.BusinessLogic.Services.DbServices.SubscriptionServ
             var show = await GetPodcastShow(podcastShowId);
             if (show == null)
             {
-                return (false, $"Podcast show with Id: {podcastShowId} is not found {insideMessage}");
+                return (false, $"Podcast show is not found {insideMessage}");
             }
             if (show.PodcastChannelId != null)
             {
@@ -2814,20 +2831,20 @@ namespace SubscriptionService.BusinessLogic.Services.DbServices.SubscriptionServ
             }
             if (show.DeletedAt != null)
             {
-                return (false, $"Podcast show with Id: {podcastShowId} has already been deleted {insideMessage}");
+                return (false, $"Podcast show has already been deleted {insideMessage}");
             }
             var showStatusId = show.PodcastShowStatusTrackings.OrderByDescending(ss => ss.CreatedAt).Select(ss => ss.PodcastShowStatusId).First();
             if (showStatusId == (int)PodcastShowStatusEnum.Draft)
             {
-                return (false, $"Podcast show with Id: {podcastShowId} is in Draft status {insideMessage}");
+                return (false, $"Podcast show is in Draft status {insideMessage}");
             }
             if (showStatusId == (int)PodcastShowStatusEnum.TakenDown)
             {
-                return (false, $"Podcast show with Id: {podcastShowId} has been taken down {insideMessage}");
+                return (false, $"Podcast show has been taken down {insideMessage}");
             }
             if (showStatusId == (int)PodcastShowStatusEnum.Removed)
             {
-                return (false, $"Podcast show with Id: {podcastShowId} has been removed {insideMessage}");
+                return (false, $"Podcast show has been removed {insideMessage}");
             }
             return (true, string.Empty);
         }
@@ -2841,16 +2858,16 @@ namespace SubscriptionService.BusinessLogic.Services.DbServices.SubscriptionServ
             var channel = await GetPodcastChannel(podcastChannelId);
             if (channel == null)
             {
-                return (false, $"Podcast channel with Id: {podcastChannelId} is not found {insideMessage}");
+                return (false, $"Podcast channel is not found {insideMessage}");
             }
             if (channel.DeletedAt != null)
             {
-                return (false, $"Podcast channel with Id: {podcastChannelId} has already been deleted {insideMessage}");
+                return (false, $"Podcast channel has already been deleted {insideMessage}");
             }
             var channelStatusId = channel.PodcastChannelStatusTrackings.OrderByDescending(cs => cs.CreatedAt).Select(cs => cs.PodcastChannelStatusId).First();
             if (channelStatusId == (int)PodcastChannelStatusEnum.Unpublished)
             {
-                return (false, $"Podcast channel with Id: {podcastChannelId} is in Draft status {insideMessage}");
+                return (false, $"Podcast channel is in Unpublished status {insideMessage}");
             }
             return (true, string.Empty);
         }
